fix: move traits::circuit_supernova inside supernova::circuit
