+++ src/traits/mod.rs
@@ -160,6 +160,5 @@ impl<G: Group, T: TranscriptReprTrait<G>> TranscriptReprTrait<G> for &[T] {
 }
 
 pub mod circuit;
-pub mod circuit_supernova;
 pub mod evaluation;
 pub mod snark;
